Cars_rental/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-taggit==3.0.0
Pillow==10.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.4.4
tzdata==2023.3
